+++ libs/components/text-editor/projects/text-editor/src/plugin-resources/code-examples/rich-text-display/rich-text-display-demo.module.ts
@@ -1,0 +1,29 @@
+import {
+  NgModule
+} from '@angular/core';
+
+import {
+  CommonModule
+} from '@angular/common';
+
+import {
+  SkyRichTextDisplayModule
+} from '@skyux/text-editor';
+
+import {
+  RichTextDisplayDemoComponent
+} from './rich-text-display-demo.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    SkyRichTextDisplayModule
+  ],
+  declarations: [
+    RichTextDisplayDemoComponent
+  ],
+  exports: [
+    RichTextDisplayDemoComponent
+  ]
+})
+export class RichTextDisplayDemoModule { }