+++ libs/components/text-editor/src/app/public/public_api.ts
@@ -1,9 +1,0 @@
-export * from './modules/rich-text-display/rich-text-display.module';
-
-export * from './modules/text-editor/types/font-state';
-export * from './modules/text-editor/types/menu-type';
-export * from './modules/text-editor/types/style-state';
-export * from './modules/text-editor/types/text-editor-merge-field';
-export * from './modules/text-editor/types/toolbar-action-type';
-
-export * from './modules/text-editor/text-editor.module';