+++ libs/components/text-editor/CHANGELOG.md
@@ -1,12 +1,33 @@
+# 5.0.0-beta.4 (2021-09-27)
+
+- Migrated to Angular CLI. [#23](https://github.com/blackbaud/skyux-text-editor/pull/23)
+
 # 4.0.0-beta.2 (2021-09-15)
 
 - Added the ability to disable the text editor. [#21](https://github.com/blackbaud/skyux-text-editor/pull/21)
+
+# 5.0.0-beta.3 (2021-07-29)
+
+- Added bug fixes and features from the `master` branch. [#19](https://github.com/blackbaud/skyux-text-editor/pull/19)
 
 # 4.0.0-beta.1 (2021-07-29)
 
 - Updated the toolbar to follow accessibility best practices. [#15](https://github.com/blackbaud/skyux-text-editor/pull/15)
 - Updated the toolbar icons for modern theme. [#16](https://github.com/blackbaud/skyux-text-editor/pull/16)
 - Fixed the text editor component to retain selection and cursor position when modifying the formatting. [#10](https://github.com/blackbaud/skyux-text-editor/pull/10)
+
+# 5.0.0-beta.2 (2021-07-15)
+
+- Added support for `5.0.0-beta.*` versions of SKY UX component libraries. [#12](https://github.com/blackbaud/skyux-text-editor/pull/12)
+
+# 5.0.0-beta.1 (2021-06-25)
+
+- Fixed the Angular peer dependencies to only support Angular 9 and 12. [#7](https://github.com/blackbaud/skyux-text-editor/pull/7)
+
+# 5.0.0-beta.0 (2021-06-24)
+
+- Initial `5.0.0-beta` release.
+- Added support for `@angular/core@^12`. [#3](https://github.com/blackbaud/skyux-text-editor/pull/3)
 
 # 4.0.0-beta.0 (2021-06-18)